Bug fixes + a few new items
- added water
- added tooltip to the UI to indicate the currently equipped item
- particle effects for liquids are more customizable
- fixed items not fully deleting when used (i.e. potions)
- improved inventory handling when calling deleteAllObjects, there is
now an option to ignore any items in the inventory
- fixed some projectiles having an inverted knockback effect
- fixed AI not following player while jumping
- fixed player sinking into the ends of platforms
- projectiles disappear when going through certain liquids
- grappling hook is more accurate
- player is not affected by gravity when being pulled by grappling hook
- only one hook can be shot at a time
- if selected slot is full, items are added to the next available slot
- added chests (still buggy)
- crafting menu items are properly aligned
- separated player's range hitbox into line of sight and range
- added a few demo levels
- the Timer class is now automated, any objects that create an instance
of the Timer class do not have to manually update it unless needed
- fixed bug where items do not delete after being used in a crafting
recipe
- added ammo counts to the hotbar (setting the ammo count to -1 will
give the item infinite ammo and the number will not be rendered in the
hotbar)

+++ database/ObjectList.java
@@ -1,12 +1,13 @@
 package database;
 
 import item.Item;
-import particles.Particle;
+import particle.Particle;
 import liquid.Liquid;
 import platform.NormalPlatform;
 import player.Player;
 import powerup.PowerUp;
 import enemy.AI;
+import engine.Timer;
 
 import java.util.LinkedList;
 import org.newdawn.slick.Graphics;
@@ -22,11 +23,20 @@ public class ObjectList {
     public static LinkedList<Object> platforms = new LinkedList<Object>();
     public static LinkedList<Object> particles = new LinkedList<Object>();
     public static LinkedList<Object> liquids = new LinkedList<Object>();
+    public static LinkedList<Object> timers = new LinkedList<Object>();
     public static Player player = new Player(0, 0);
 
     public static void updateAllObjects() {
 
         player.update();
+        
+       try {
+            for (int i = 0; i <= ObjectList.timers.size(); i++) {
+                ((Timer) ObjectList.timers.get(i)).update();
+
+            }
+        } catch (Exception e) {
+        }
 
         try {
             for (int i = 0; i <= ObjectList.items.size(); i++) {
@@ -62,7 +72,7 @@ public class ObjectList {
 
         try {
             for (int o = 0; o <= ObjectList.objects.size(); o++) {
-                ((object.Object) ObjectList.objects.get(o)).update();
+                ((levelobject.Level_Object) ObjectList.objects.get(o)).update();
 
             }
         } catch (Exception e) {
@@ -91,7 +101,7 @@ public class ObjectList {
         //render all objects
         try {
             for (int o = 0; o <= ObjectList.objects.size(); o++) {
-                ((object.Object) ObjectList.objects.get(o)).draw(g);
+                ((levelobject.Level_Object) ObjectList.objects.get(o)).draw(g);
 
             }
         } catch (Exception e) {
@@ -170,9 +180,10 @@ public class ObjectList {
                     if (items.get(i) != null) {
                         if (destroyInventory == true) {
                             ((Item)items.get(i)).delete();
-                            Inventory.deleteItem(items.get(i));
                         } else {
-                            ((Item)items.get(i)).delete();
+                            if (Inventory.contains(items.get(i)) == false) {
+                                ((Item)items.get(i)).delete();
+                            }                            
                         }
                     }
 
@@ -195,7 +206,7 @@ public class ObjectList {
             for (int i = 0; i <= objects.size(); i++) {
                 try {
                     if (objects.get(i) != null) {
-                        ((object.Object) objects.get(i)).delete();
+                        ((levelobject.Level_Object) objects.get(i)).delete();
                     }
                 } catch (Exception e) {                 
                 }
